test: Stop sending perf benchmarks telemetry to App Insights (#25088)
## Description

At some point we had the idea of try using AppInsights to store perf
benchmarks telemetry to be able to build better workflows and dashboards
around it (e.g. make it possible to everyone to set thresholds for
benchmarks). This never materialized and I don't think it ever will, so
this PR removes the pipeline steps that send perf benchmarks telemetry
to app insights as a pre-requisite to eventually removing the relevant
code in our internal infrastructure.