Update to chai 5.2.1 (#25087)
## Description

Some of our tests use chai for asserts (most just use node's asserts
module).

This updates the chai version to 5, if the release group is tooling and
does not need CommonJS support.

A couple packages also had unused dependencies on chai which were
removed.

Some packages in the client release group are still using chai, and as
they need to support CommonJS they are not updated.
 
For the cases which did update, or removed the dependency, removes a
dependency on the old version of loupe so that throws exceptions on
startup which is annoying.