flub fix: Dockerfile policy check fails on Windows (#25100)
## Description

Fixed an issue with flub that only shows up on Windows. The paths we are
checking in Dockerfiles as part of the policy check use foreslashes, but
we build it using paths from the filesystem. This change normalizes the
path to use foreslashes.

Note that we need to make a pass over the other places that build
filepaths to check against file contents. This will be a separate work
item.