Only apply locks to odsp and frs stages in perf benchmarks and stress tests (#25107)
## Description

Today all stages of the perf benchmarks pipeline and the stress tests
pipeline have to wait for the corresponding stage in other runs to
finish before starting itself. This is not necessary in all stages, only
in the ones that we know could stress external services. This PR makes
it so only the stages against ODSP and FRS have that behavior.

Same idea as https://github.com/microsoft/FluidFramework/pull/24942
which did the same thing for end-to-end tests.