build: Add rerun attempt number to artifacts published by the perf benchmarks pipeline (#25110)
## Description

Minor quality of life improvement. Today re-running a job of the
performance benchmarks pipeline runs the risk of failing when trying to
publish the artifact with the results if the previous attempt managed to
get to that point and published one already. This PR adds the attempt
number to the artifact name so each re-run publishes a separate artifact
and they don't complain about it already existing.

Confirmed that it works as expected with a test run of the pipeline:

<img width="584" height="319" alt="image"
src="https://github.com/user-attachments/assets/20d772bb-6818-47e7-820a-999aa13dbe64"
/>