fix(tree): boundary condition for usage of new DetachedFieldIndex format (#25117)
## Description

Amends the logic that decides when to use the new `DetachedFieldIndex`
format. Before this change `CodecWriteOptions.oldestCompatibleClient`
had to be set to a value greater than `FluidClientVersion.v2_52`. After
this change, it has to be set to a value equal to or greater than
`FluidClientVersion.v2_52`.

## Breaking Changes

The new format will now be used in situations where it wasn't before.
This should not be a problem since no public release has been made since
the new format was introduced.