(tree) Renamed treeShaper / fieldShaper and related concepts (#25109)
Renamed some of the objects / types related to tree shaper and field
shaper. The goal is to make these concepts less confusing and easier to
understand, especially because the terms `node` and `tree`, and `shape`
and `encoder` are used somewhat interchangeably. This also complements
https://github.com/microsoft/FluidFramework/pull/25058 which renames
other parts of the encoding logic with a similar goal.

The overall theme of this change is:
1. Update some of the concepts from using the term `shape` to `encoders`
since their primary objective is encoding. The fact that these encoders
are shape too is an implementation detail.
2. Update these concepts to use the term `node` instead of `tree`. While
`tree` is an appropriate name when the node is considered along with its
children but `node` seems more consistent with rest of the system.

Here are the renames:
- Interface `TreeShaper` -> `NodeEncodeBuilder` since it has properties
that help build a node encoder from given paramteters.
- In the interface, property `shapeFromTree` -> `nodeEncoderFromSchema`
because it returns a node encoder which is a shape as well but that is
an implementation detail.
- Interface `FieldShaper` -> `FieldEncodeBuilder` since it has
properties that help build a field encoder from given paramteters.
- In the interface, property `shapeFromField` ->
`fieldEncoderFromSchema` because it returns a node encoder which is a
shape as well but that is an implementation detail.
- Type `FieldShapePolicy` -> `FieldEncoderPolicy` since it defines a
type that can be used to create a field encoder from the given params.
- Type `TreeShapePolicy` -> `NodeEncoderPolicy` since it defines a type
that can be used to create a node encoder from the given params.
- Renames of variables and parameters related to the above.


[AB#45293](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/45293)