build(build-tools): Update build-tools dev dependency (#25127)
## Description

Update the version of build-tools installed from npm to build the
in-repo version of build-tools, to the latest release. This gets rid of
the transitive dependency on `parse-git-config` which is the reason for
a Component Goverannce alert.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).


[AB#35181](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/35181)