(compat) Move Driver layer compat validation to Container creation time (#25120)
#24462 and #24855 added layer compat validation logic across the Loader
/ Driver layer. In the driver layer, it added the `ILayerCompatDetails`
on document service and the Loader layer validates that when Container
creates a document service.
However, we should be able to detect layer incompatibility much sooner
and fail fast. This PR does that by moving the `ILayerCompatDetails`
from document service to document service factory. The validation in
Loader layer now happens in the Container constructor.

[AB#45406](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/45406)