Initial applyStashedOp implementation in Shared Array (#25140)
This PR consists in a first implementation to applyStashedOp in shared
array, considering mainly basic ops insert, delete and move.

Implementation consisted in recreating what the DDS would do while
processing each type of operation being treated as remote. The reason
for this is because shared array ops are not easily reproducible from
the type of metadata they have. For example, map set op contains the key
and the value fields, which is exactly what set parameters are and is
not the case for shared array insert op, which needs index and value as
well but op contains entryId, value and insertAfterEntryId. That
translation is already done in op remote processing so currently just
reusing that logic for each operation type.

Adding a couple e2e tests to verify implementation, being conscious
those are the minimum possible testing we could add and once local
server stress tests are enabled we could start hitting bugs.

toggle and toggleMove ops require a bit more nuance to simulate so
adding more tests for those in a subsequent PR