docs(protocol-definitions): Bump release level of legacy APIs (#25179)
We are in the process of migrating to a new "legacy" tagging spec with
the following tags -> exports mapping:

- `@legacy @beta` -> `/legacy`
- `@legacy @alpha` -> `/legacy/alpha`

In preparation for this update, all `@legacy @alpha` APIs in this
package have been promoted to `@legacy @beta`.

Note that our current tooling maps _all_ `@legacy` APIs to `/legacy`, so
this change is safe to make in advance of the tooling changes, and is in
fact a necessary prerequisite.


[AB#45987](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/45987)