Fix(tree): rebase over dependent change of conflicted transaction (#25191)
## Description

This PR fixes a document-corrupting bug. Before this PR, making
concurrent edits to an array or a sequence could lead to a firing of
assert `0x8a2` if the following conditions were met:
* Some edit `e1` was a transaction with a constraint that turned out to
be violated by edits concurrent to (and sequenced before) `e1`
* Some edit `e2` was dependent on `e1` (from before the violation of its
constraint)
* Some edit `e3` was concurrent to and sequenced after both `e1` and
`e2`
* `e3` was either rebased over or the revert of some other edit `e0`
that predated `e1`, `e2`, and `e3`.
* `e0` and `e2` made edits to the same gap (i.e., in the same space
between nodes) in the sequence/array.
(See the test added by this PR for an example)

This was due to the fact that our current cell ordering scheme depends
on acquiring tombstones when rebasing over a change with conflicts, but
our implementation of rebase and compose did not allow this.

In the future, we would rather not require rebasing over conflicted
changes has an effect on the rebased changes. See
https://dev.azure.com/fluidframework/internal/_workitems/edit/46104.

## Breaking Changes

None