Restrict typing of annotated types (#25197)
## Description

Restrict annotated allowed types to only be allowed to be made be FF
APIs (currently just SchemaFactoryAlpha.staged) by annotating them with
`@sealed`.

Also mark implicit schema types is `@input` so adding more options to
them in the future is not a breaking change.

This makes it possible to stabilize that AnotatedAllowedTypes are a
thing, without having to stabilize their exact format, or any particular
annotation (like staged schema)

## Breaking Changes

Existing code using these alpha APIs may need to be updated if it is
violating the restrictions of the new tags.