feat(directory): Directory rollback across remote ops (Part 1 -  storage ops only) (#25126)
## Description

This PR is a prototype for rollback across remote ops for directory.
This PR is based off of
https://github.com/microsoft/FluidFramework/pull/25070 and implements
rollback for storage ops (set, delete, clear).

Note: Rollback acrros remote ops for subdirectories is **not**
implemented in this PR and will be included in a follow up.

## Misc

[AB#43738](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/43738)
Prototype: https://github.com/microsoft/FluidFramework/pull/25092
Will fix
[AB#38522](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/38522)

---------

Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>