test(matrix, tree): Update matrix/table perf tests to opt into event listeners and consistently opt into undo/redo  (#25195)
Updates performance tests to bind event listeners to the matrix/table,
and to consistently opt into undo/redo management. We are using these
tests to compare SharedTree-based table performance to that of
SharedMatrix. Since the real-world scenarios we care about will leverage
both event listeners and undo/redo, we need this for our tests to be
representative.

Also substantially refactors test helpers to allow more code sharing (to
prevent having to make policy changes like this in more places).


[AB#46138](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/46138)