fix(Test logs): Avoid squashing 'details' logging property (#25202)
In our test logs, every log has the `details` property set to the same
thing, something like `{ "displayName": "RealStressService"}` (depending
on the specific test). This is because the mocha hooks overwrite the
`details` property.

The fix here is to put that `displayName` property (and any
other props added via ENV) in the logs under `testEnvProps`, leaving
`details` for whatever the actual code said.