Refactor SharedTree initialize logic (#25208)
## Description

Add some missing validation, move some sequence root handling logic to a
test utility, reduce coupling between utilities in schematizeTree.ts,
and make more of the relevant code use TreeChunk to match the underling
storage.

Additionally clarify in the docs when the callback is run as this is
both important now to allow use of forest.chunkTree (which needs the
schema to have been set) but also for the future when moving to an
create then attach model.