Sensitivity label parsing fix in ODSP join session call (#25186)
## Description

Fixes parsing of sensitivity label information in the ODSP join session
response. So this was double-parsed, leading to runtime errors when
sensitivity label information was actually present.

In practice, sensitivity label information hasn't been present, because
it wasn't rolled out ODSP side and we didn't sent the
`Return-Sensitivity-Labels` Prefer header. We're now setting this
header, but have control over it via a feature gate. Once ODSP rollout
is done, we can test and roll out on our end, and eventually delete the
feature gate again.

## Breaking Changes

This breaks the legacy/alpha API of sensitivityLabelsInfo in
ISocketStorageDiscovery, switching from string to a more specific object
type. At runtime, when sensitivity label information had been present,
this would have been an object all along. However, in production this as
not been present, and shouldn't have been used by any consumer. So this
makes sensitivityLabelsInfo correct and usable for the first time.