Fix restart-from-pending to exit staging mode and correct savedOps clearing logic (#25180)
Previously, the restart-from-pending operation could select a client
that was still in staging mode in the generator. A better approach is to
check if `sourceClient` is in staging mode and call `exitStagingMode` in
the reducer if so.

Additionally, there was a bug in replayPendingMessages where savedOps
was cleared even when the op was not staged (stages: false). This has
been updated so that savedOps is only cleared when
committingStagedBatches is false. This change fixed most of the failing
seeds.