test(tree): Add end-to-end test for revert constraint (#25224)
## Description

This PR updates an existing test. The existing test demonstrated that
revert constraints work in a scenario involving edits on a local branch.
The updated test demonstrates that revert constraints work in a scenario
that involves peers communicating across encoding and decoding
boundaries.

The motivation for this test was a fear that, because
`ModularChangeFamily` codecs ignore revert constraints, the constraint
would only be respected on the peer that emitted the edit with the
revert constraint. This is not the case because revert constraints are
turned into "normal" constraints when reverting, and it's the normal
constraint that needs to be sent across peers. Note this relies on the
assumption that the revert commit generated from the commit with the
revert constraint is only ever generated by the peer who authored the
commit with the revert constraint. IOW, that a peer can't undo/redo
another peer's changes.

## Breaking Changes

None