Support instanceof TreeNode (#25226)
## Description

Support `instanceof TreeNode`.

In the past there was some reluctance to support this. Users however
have tried to use it and noted it not being supported. Opinions seem to
be on the side of supporting it now.

This API can't trivially be stabilized incrementally using release tags
since type and non-type exports are required to have matching tags (and
docs). It could be aliased to a different name, and that exported as
TreeNodeAlpha. Generally, though this seems like a low risk change as
this has been working an supported internally in the package for a long
time, so going straight to public seems reasonable.

Note that API reports do not capture if something it type only exported,
so it is a defect in API reporting that this does not show up as a
change to all API surfaces for both tree and fluid-framework.