test(matrix): Refactor matrix perf tests to make memory and execution time tests use the same configurations and structure (#25216)
We have 2 sets of tests that are intended to be the same but are run
independently to measure execution time vs memory consumption. The sets
of tests were structured very differently from one another but are
intended to test the same scenarios.

This PR updates the tests to be more structurally similar and share more
infrastructure.

In a future PR, I plan to update the overall test pattern to declare the
set of test scenarios once and run execution time and memory consumption
tests against the same inputs. This will make future maintenance
significantly easier.

Additionally, similar changes will be made to the corresponding
SharedTree table tests.


[AB#46340](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/46340)