Fix webpack-fluid-loader's incremental build (#25241)
## Description

Our incremental build failed to pick up changes to webpack-fluid-loader
when a build was run from an example using webpack-fluid-loader.
Example:

1. Build repo, run inventory-app
2. Make a change to `examples\utils\webpack-fluid-loader\src\loader.ts`
3. Run `npm run build` from the inventory app
4. Run `npm run start` from the inventory app

This flow should let you see the change you made to `loader.ts`, but it
did not. This is because webpack-fluid-loader injects a script tag
referencing a code bundle it produces via its webpack script, but the
`npm run build` in step 3 would not rerun that bundle step due to no
apparent dependency in our fluid-build config.

This change adjusts webpack-fluid-loader's fluid-build config to
reasonably declare this dependency.

Co-authored-by: Abram Sanderson <absander@microsoft.com>