(tree): Incremental summarization of chunked forest (#24949)
## Description

[AB#1263](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1263)
has an overall description and link to a design document.

This PR adds incremental summarization support to chunked forest.
Currently, the chunked forest data is encoded and added to a single blob
in the summary tree. This change breaks it down into mutiple subtrees
where each subtree can be incrementally summarized, i.e., if such a
subtree doesn't change between summaries, a summary handle will be added
to the summary instead of the entire subtree.
The schema will support marking properties as incrementally summarizable
(TODO - not in this PR). Any chunks in the fields corresponding to this
property will be the boundary at which incrementally summarization will
work. Basically, if this property changes or anything within it changes
(for objects and arrays), its chunks will be re-summarized, or else, a
summary handle will be used.

To achieve this, this PR makes the following changes:
- Added a new encoded format called `EncodedIncrementalChunkShape` to
`EncodedChunkShape` for chunks whose data will be incremental
summarized.
- Added a new shape called `IncrementalChunkShape` which represents
chunks whose data can be incrementally summarized, aka, incremental
chunks.
- Each incremental chunk is assigned a `ChunkReferenceId` which uniquely
identifies it under its parent field.
- Added a new field encoder called `incrementalFieldEncoder` which
encodes its chunks as `IncrementalChunkShape`. The `ChunkReferenceId`s
of all chunks in a field are stored as `EncodedNestedArrayShape` in the
encoded data.
- Added a new decoder called `IncrementalChunkDecoder` which can decode
an incremental chunk.
- Added `IncrementalEncoder` and `IncrementalDecoder` interfaces which
has properties that help with encoding and decoding of incremental
chunks. These are added to the `FieldBatchEncodingContext`.
- Added a `ForestIncrementalSummaryBuilder` which tracks chunks across
summaries and generates summary tree or summary handles for them
depending on whether they change between summaries.
- The forest summary tree consists of a blob called `ForestTree` at the
root that contains the data for the forest. If there are any incremental
fields in the forest, an array of chunk reference ids will be encoded
for its chunks in the data.
- Fore each incremental chunk, a separate summary tree will be added
with its reference id as the key for the tree.
- The data for the chunk will be added to a blob called `contents` under
the tree. If the chunk has any incremental fields, an array of chunk
reference ids will be encoded for its chunks in the data.
- The same format is repeated until all incremental fields have been
summarized.
- Note that currently, an incremental field will have exactly one chunk.
This may change in the future where the data may be encoded into
multiple chunks.


[AB#14308](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/14308)
[AB#41861](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41861)
[AB#41863](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41863)
[AB#41864](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41864)