(tree) Added validation that forest is loaded from a valid snapshot (#25275)
Currently, if `ForestSummarizer` is loaded from a malformed snapshot --
one that doesn't contain a summary blob for the top-level forest
contents -- it will essentially load an empty forest leading to possible
data loss and corruption.

There must always be a summary blob containing the top-level forest
contents under a specific key. This change asserts this which will
ensure that malformed snapshot would cause the forest load to fail.

 

[AB#46751](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/46751)