docs(local-driver): Promote `LocalDocumentServiceFactory` to `@legacy @beta` (#25290)
This type was not bumped previously (#25256) because of tag
incompatibility with a type it imports from
`@fluidframework/server-local-server`.

The server packages are currently configured to use release tags, but
their exports do not respect them for API access control. For this
reason, their APIs can be considered implicitly public, regardless of
tagging. In the future, we need to update the server packages to either
a. not use release tags
b. match client's release tag -> export path scheme (this would be a
breaking change and therefore require a major version bump to those
packages).

Once those packages have been updated to follow the same release tag /
export schema as the client packages, then we can resume linting tag
compatibility across the client/server boundary. Until then, we will
need to add exceptions.