Remove isSessionActive check for broadcasting signals (#25246)
## Description

This PR aims to fix the broadcast-signal endpoint in alfred, where all
requests were responding with a 410. The solution basically removes the
`isSessionActive` check, as it doesn't pose a risk to the logic in
general and `redis-emitter` will still be able to handle the emission
without errors if the session is active.

## Reviewer Guidance

This is a simple PR, removing one check that was done initially before
broadcasting a signal. The removal of this check does not affect
functionality or break the API. The check was removed as tests from PP
resulted in the API always failing the check.