test(matrix): Unify memory and runtime benchmark test suites (#25287)
We previously had the same test scenarios duplicated across two
benchmark test suites: 1 for memory benchmarking, and one for execution
time benchmarking. This PR unifies the suite of test scenarios into a
single file and updates the memory and runtime benchmark suites to
leverage the single set of test declarations.


[AB#46340](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/46340)