DataStoreRuntime: Remove redundent attachGraph in getAttachmentSummary (#25310)
Per the comment and the code, the call the attachGraph on the contexts
should be redundant. I also believe we have sufficient test coverage for
this case given our attach lifecycle tests, binding tests, and local
server stress. If reviewers believe there isn't sufficient coverage for
this case I can add more tests, but i believe it is well covered.


This change is a pre-cursor change to returning detached but referenced
datastore in pending local state while under staging mode. There are
additional issues that block that specific case, so we can't test it
yet. This causes problems in that staging mode specifically because we
defer attach, and don't support sending attach ops while in staging mode
which the triggering of attachGraph attempts to do.