Add readme to tree tests (#25362)
## Description

Add readme to tree tests, and note some details for --exit issue.

Since we don't have consensus on exactly how to organize and name
integration tests, I figured I'd leave details on that for a future
change, which can also include adjusting some code to follow the pattern
as an example.