feat(task-manager): Add rollback support for TaskManager (#25388)
## Description

This PR adds support for rollback for TaskManager (including across
remote ops).

## Implementation

Since TaskManager is a fully consensus-based DDS, we don't need to do
much to support rollback. We mainly do two things:
1. Ensure the op we are trying to rollback exists in the pending data
(and cleanup pending data since we will no longer process the op).
2. Use a new `rollbackWatcher` event emitter to ensure we properly
resolve promises that were waiting on the pending op to be ack'd.

Note: There are a few miscellaneous changes in this PR that are not
directly related to rollback, but were exposed when enabling local
server stress tests (and fixed as part of this PR).

## Testing

This PR also adds additional testing:
1. Added rollback specific testing in `taskManager.rollback.spec.ts`
4. Enabled rollback in TaskManager fuzz testing
5. Enabled local server stress tests to use TaskManager

## Misc

[AB#44704](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/44704)

---------

Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>