Use TreeNodeSchemaClass in ValidateRecursiveSchema (#25415)
## Description

All recursive schema types are instances of TreeNodeSchemaClass. This
was documented on ValidateRecursiveSchema's implementation, but not used
in its typing.

Since TreeNodeSchema is implemented as a union of TreeNodeSchemaClass
and TreeNodeSchemaNonClass, this change makes ValidateRecursiveSchema
slightly stricter, but in a way all recursive schema will pass, and also
simpler, since TypeScript no longer has to infer which kind of schema
the input is.

In practice, this should not impact any users of the API, buts just now
a little more robust and gives slightly better errors.