[test/data-migration]: Generic approach to MigrationDataObject that supports multiple internal Fluid Data Models (#25382)
Main idea - the MigrationDataObject separates the external data
interface (i.e. getting/setting arbitrary keys, maybe a SharedTree
instance) from the internal fluid data model used to implement it (i.e.
which DDSes are actually there in the Runtime).

To do this it takes in a `TUniversalView` type parameter (the external
data interface), and a list of one or more "Model Descriptors" which can
identify whether the DataStoreRuntime matches that model (similar to
`refreshRoot` in the base branch), as well as how to initialize a new
DataObject with that model. Each Model Descriptor yields a model that
fits into `TUniversalView`.

The simplest way to specify `TUniversalView` is to just union all the
individual model types. (The example below makes a different choice)

### Updating DataObject and TreeDataObject

This PR also illustrates updating DataObject and TreeDataObject to use
MigrationDataObject. We could deprecate those if we get
MigrationDataObject to a place we like, and existing subclasses of
DataObject could switch to MigrationDataObject directly. Then they'd be
ready if/when a migration is needed. Maybe we simply put the new
capabilities on PureDataObject instead of introducing a new class.

### Example

See `demo.ts`.

The Model Descriptor for the classic "root SharedDirectory" DataObject
checks for a channel with ID "root" that's a SharedDirectory, whereas
the Descriptor for the "root SharedTree" DataObject checks for
"root-tree" channel that's a SharedTree.

The `TUniversalView` here is this abstraction, taken from the prototype
component code:

```typescript
  readonly getArbitraryKey: (key: string) => string | boolean | undefined;
  readonly setArbitraryKey: (key: string, value: string | boolean) => void;
  readonly deleteArbitraryKey: (key: string) => void;
  readonly getRoot: () => {
    isDirectory: true,
    root: ISharedDirectory
  } | {
    isDirectory: false,
    root: ITree,
  };
```

## Reviewer Guidance

There are plenty of todos and some gaps. Some listed below. Any and all
feedback is welcome, but focusing on higher-level is more helpful in the
short term.

### Gaps / open questions 
- We need something like an "evacuateModel" function on ModelDescriptor
to clean up unreferenced channels after migration
- API / experience around how you'd roll out a new ModelDescriptor and
stage the migration is not fleshed out.
- Reminder this is targeting a WIP feature branch, so the rest of the
migration flow is PoC quality as well

Thanks for taking a look!