improvement(runtime): Add fail fast mechanism for runtime options that require explicitSchemaControl (#25430)
## Description

This PR adds a fail fast mechanism that will throw a `UsageError` if a
runtime option that requires `explicitSchemaControl` is enabled without
enabling `explicitSchemaControl` as well.

To avoid disruption for existing customers, we will only apply this to
runtime options that were added after `explicitSchemaControl` was
introduced. Currently, this only applies to `createBlobPayloadPending`,
but going forward FF devs will be required to add new runtime options to
`runtimeOptionKeysThatRequireExplicitSchemaControl` (the build will not
compile otherwise). This will enforce this fail fast mechanism for any
new runtime options going forward.

## Misc

[AB#48150](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/48150)

---------

Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>