fix(tree): Buffer events during multi-op node edits which are intended to be atomic (#25346)
Add package-internal functionality to allow pausing and buffering
TreeNode events. This can be used to avoid user event spam in controlled
scenarios where we know it is safe to do so. This functionality is not
exposed to the user, as the behavior could be dangerous when not
considered very carefully.

Updates Table schema operations to pause events during operations that
are composed of multiple tree-level edits. This (when combined with
transactions) gives the appearance of atomicity for these operations,
which is the desired user-facing behavior.

Also mitigates an existing bug
([AB#47457](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/47457))
that caused 2 tree change events to be emitted when moving a child
within or between unhydrated array node(s).


[AB#47668](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/47668)